calcNonrenGroundwatUse

Calculates current non-renewable (fossil) groundwater use as negative difference between water availability from renewable sources and human water use (consumption and withdrawal). The water use in the initialization year is fixed in the future as non-renewable groundwater.

Usage

calcNonrenGroundwatUse(
  output,
  lpjml,
  climatetype,
  transDistGW = 100,
  multicropping = "TRUE:actual:irrig_crop",
  selectyears,
  iniyear
)

Arguments

output

"total": total groundwater use (non-agricultural and agricultural), "nonAg": groundwater use in non-agricultural sector (industry, domestic, electricity), "comAg": groundwater use in agricultural sector (currently irrigated area)

lpjml

LPJmL version used

climatetype

Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"

transDistGW

Water transport distance allowed to fulfill locally unfulfilled water demand by surrounding cell water availability to determine missing water that is tagged as non-renewable groundwater

multicropping

Multicropping activated (TRUE) or not (FALSE) and Multiple Cropping Suitability mask selected (mask can be: "none": no mask applied (only for development purposes) "actual:total": currently multicropped areas calculated from total harvested areas and total physical areas per cell from LandInG "actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific), "actual:irrig_crop" (crop- and irrigation-specific) "total" "potential:endogenous": potentially multicropped areas given temperature and productivity limits "potential:exogenous": potentially multicropped areas given GAEZ suitability classification) (e.g. TRUE:actual:total; TRUE:none; FALSE)

selectyears

Years to be returned

iniyear

Initialization year

Value

cellular magpie object

Author

Felicitas Beier

Examples

if (FALSE) { # \dontrun{
calcOutput("NonrenGroundwatUse", aggregate = FALSE)
} # }